chore: configure and build Dev Container for flutter development
This commit adds a Dev Container configuration to allow building and
testing the Android Flutter application without additional setup on the
host.
It also presents a method of using the Dev Container in a cloud
workspace (GitHub Codespaces) while still being able to push the builds
to a connected Android device.